refactor: dbの位置を変更
