+++ internal/infra/postgres/user.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgxpool"
-	"github.com/saitamau-maximum/maxicloud/internal/db"
 	"github.com/saitamau-maximum/maxicloud/internal/domain"
+	"github.com/saitamau-maximum/maxicloud/internal/infra/postgres/db"
 )
 
 type userRepository struct {
