+++ sqlc.yaml
@@ -6,6 +6,6 @@ sql:
     gen:
       go:
         package: "db"
-        out: "internal/db"
+        out: "internal/infra/postgres/db"
         sql_package: "pgx/v5"
         emit_pointers_for_null_types: true
